chore(occt-wasm): extract distance/surfaceCurvature into measureOps
Moves the complex measure paths (`distance`, `surfaceCurvature`) and
their helpers (`collectDistanceSamples`, `computePrincipalDirections`,
`surfaceDerivatives`, `eigenvector2x2`, `liftAndNormalize`,
`degenerateOrthoFrame`, `pointAt`) out of `occtWasmAdapter.ts` and into
`measureOps.ts` alongside the simple wrappers that already lived there.

Adapter line count: 2288 → 1956 (-332). Adapter no longer carries
module-scope helpers other than `wrapKernelExceptions` and
`notImplemented`.
